+++ src/constants/index.ts
@@ -1,0 +1,2 @@
+export const TABLET_BREAKPOINT = 768;
+export const MOBILE_BREAKPOINT = 576;
